h1st/django/requirements.txt
# HUMAN-FIRST AI
H1st >= 2020.12


# DJANGO
Django >= 3.1.8, < 3.2

# Django Polymorphic Models
Django-Polymorphic >= 3.0.0

# Django Admin Themes
Django-Grappelli >= 2.14.3
Django-Jazzmin >= 2.4.5

# Django-AutoComplete-Light & related
Django-AutoComplete-Light >= 3.8.2
Django-Generic-M2M >= 0.3.1
Django-GM2M >= 1.1.1
Django-QuerySetSequence >= 0.14
Django-Tagging >= 0.5.0
Django-Taggit >= 1.3.0

# Django REST Framework
DjangoRESTFramework >= 3.12.4
CoreAPI >= 2.3.3
    # closely related to Django REST Framework
DjangoRESTFramework-JSONAPI >= 4.1.0

# Django Filtering
Django-Filter >= 2.4.0
DjangoRESTFramework-Filters == 1.0.0.dev2

# Django Model Query Optimization
Django-Model-Query-Graphs >= 0.0.5

# Django Query Profiling
Django-Silk >= 4.1.0

# Django Extensions & Utilities
Django-Extensions >= 3.1.2
Django-Model-Utils >= 4.1.1
Django-Storages >= 1.11.1

# Django CORS Headers
Django-CORS-Headers >= 3.7.0


# ASYNCHRONOUS SERVER GATEWAY INTERFACE (ASGI)
Daphne >= 3.0.2; python_version < '3.9'
Hypercorn >= 0.11.2
Uvicorn >= 0.13.4
HTTPTools >= 0.1.1; python_version < '3.9'
    # required by Uvicorn
Uvloop >= 0.15.2
    # required by Uvicorn


# DATABASE CONNECTORS/DRIVERS
MySQLClient >= 2.0.3; python_version < '3.9'
PsycoPG2-binary >= 2.8.6


# ML MODEL SERDE
JobLib >= 1.0.1


# OTHER
GitPython >= 3.1.14
Ruamel.YAML >= 0.17.2
Tqdm >= 4.60.0


# CLOUD
# AWS Elastic Beanstalk
AWSEBCLI >= 3.19.4
